pyproject.toml
[project]
name = "claire-directllm"
version = "0.1.0"
description = "Direct LLM baseline for cybersecurity question answering"
readme = "README.md"
authors = [
    { name = "Joel", email = "joel@metakortex.org" }
]
requires-python = ">=3.11"
dependencies = [
    "openai>=2.11.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "tiktoken>=0.12.0",
    "typer>=0.20.0",
]

# CLI is accessed via: uv run python -m claire_directllm.ask

[build-system]
requires = ["uv_build>=0.8.17,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.9",
]
